Fix edge functions logs refresh button (#35194)
* Fix edge functions logs refresh button

* add comment to move state

---------
